Refactor scala code to just make it work. Next step will be creation of separate sbt project with proper structure for easiness of usage.
