+++ CH 01 Code/ServerContext/Connection.scala
@@ -1,0 +1,7 @@
+package ServerContext {
+	object Connection {
+		def send(name: String, dataType: String, v: Float) {
+			println(s"Sended: $name $dataType $v")
+		}
+	}
+}
